requirements.txt
asgiref==3.8.1
attrs==23.2.0
black==24.4.2
click==8.1.7
Django==5.0.5
django-filter==24.2
djangorestframework==3.15.1
drf-spectacular==0.27.2
inflection==0.5.1
install==1.3.5
isort==5.13.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.18.1
ruff==0.4.3
sqlparse==0.5.0
uritemplate==4.1.1
